feat(engine): add compressOutput property to action schema and update workflows
