+++ engine/runtime/action-sink/src/file/mvt.rs
@@ -1,3 +1,4 @@
+pub(crate) mod pipeline;
 pub(crate) mod sink;
 pub(crate) mod slice;
 pub(crate) mod tags;
