+++ engine/runtime/action-sink/src/file/mvt/sink.rs
@@ -1,22 +1,12 @@
 use std::collections::HashMap;
-use std::convert::Infallible;
-use std::io::Write;
-use std::path::Path;
+use std::io::BufWriter;
+use std::io::Cursor;
 use std::str::FromStr;
+use std::sync::mpsc;
+use std::sync::mpsc::Receiver;
 use std::sync::Arc;
-use std::sync::Mutex;
 use std::vec;
 
-use flate2::{write::ZlibEncoder, Compression};
-use flatgeom::LineString2;
-use flatgeom::MultiLineString as NMultiLineString;
-use flatgeom::MultiLineString2;
-use flatgeom::MultiPolygon as NMultiPolygon;
-use flatgeom::MultiPolygon2;
-use itertools::Itertools;
-use prost::Message;
-use rayon::iter::ParallelBridge;
-use rayon::iter::ParallelIterator;
 use reearth_flow_common::uri::Uri;
 use reearth_flow_runtime::errors::BoxedError;
 use reearth_flow_runtime::event::Event;
@@ -30,16 +20,10 @@ use reearth_flow_types::Feature;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use serde_json::Value as JsonValue;
-use tinymvt::geometry::GeometryEncoder;
-use tinymvt::tag::TagsEncoder;
-use tinymvt::vector_tile;
 
-use super::slice::slice_cityobj_geoms;
-use super::tags::convert_properties;
-use super::tileid::TileIdMethod;
-use super::tiling::TileContent;
-use super::tiling::TileMetadata;
 use crate::errors::SinkError;
+
+use super::tileid::TileIdMethod;
 
 #[derive(Debug, Clone, Default)]
 pub struct MVTSinkFactory;
@@ -101,6 +85,14 @@ impl SinkFactory for MVTSinkFactory {
         let layer_name = expr_engine
             .compile(expr_layer_name.as_ref())
             .map_err(|e| SinkError::MvtWriterFactory(format!("{:?}", e)))?;
+        let compress_output = if let Some(compress_output) = &params.compress_output {
+            let compress_output = expr_engine
+                .compile(compress_output.as_ref())
+                .map_err(|e| SinkError::MvtWriterFactory(format!("{:?}", e)))?;
+            Some(compress_output)
+        } else {
+            None
+        };
 
         let sink = MVTWriter {
             global_params: with,
@@ -110,19 +102,24 @@ impl SinkFactory for MVTSinkFactory {
                 layer_name,
                 min_zoom: params.min_zoom,
                 max_zoom: params.max_zoom,
+                compress_output,
             },
+            join_handles: Vec::new(),
         };
         Ok(Box::new(sink))
     }
 }
 
-type BufferKey = (Uri, String);
+type BufferKey = (Uri, String, Option<Uri>); // (output, feature_type, compress_output)
+type JoinHandle = Arc<parking_lot::Mutex<Receiver<Result<(), SinkError>>>>;
 
 #[derive(Debug, Clone)]
 pub struct MVTWriter {
     pub(super) global_params: Option<HashMap<String, serde_json::Value>>,
     pub(super) params: MVTWriterCompiledParam,
     pub(super) buffer: HashMap<BufferKey, Vec<Feature>>,
+    #[allow(clippy::type_complexity)]
+    pub(super) join_handles: Vec<JoinHandle>,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, JsonSchema)]
@@ -132,6 +129,7 @@ pub struct MVTWriterParam {
     pub(super) layer_name: Expr,
     pub(super) min_zoom: u8,
     pub(super) max_zoom: u8,
+    pub(super) compress_output: Option<Expr>,
 }
 
 #[derive(Debug, Clone)]
@@ -140,6 +138,7 @@ pub struct MVTWriterCompiledParam {
     pub(super) layer_name: rhai::AST,
     pub(super) min_zoom: u8,
     pub(super) max_zoom: u8,
+    pub(super) compress_output: Option<rhai::AST>,
 }
 
 impl Sink for MVTWriter {
@@ -155,7 +154,8 @@ impl Sink for MVTWriter {
             )));
         };
 
-        let feature = ctx.feature;
+        let feature = &ctx.feature;
+        let context = ctx.as_context();
         match feature.geometry.value {
             geometry_types::GeometryValue::CityGmlGeometry(_) => {
                 let output = self.params.output.clone();
@@ -163,12 +163,33 @@ impl Sink for MVTWriter {
                 let path = scope
                     .eval_ast::<String>(&output)
                     .map_err(|e| SinkError::MvtWriter(format!("{:?}", e)))?;
+                let compress_output = if let Some(compress_output) = &self.params.compress_output {
+                    let compress_output = compress_output.clone();
+                    let path = scope
+                        .eval_ast::<String>(&compress_output)
+                        .map_err(|e| SinkError::MvtWriter(format!("{:?}", e)))?;
+                    Some(Uri::from_str(path.as_str())?)
+                } else {
+                    None
+                };
                 let output = Uri::from_str(path.as_str())?;
                 let layer_name = scope
                     .eval_ast::<String>(&self.params.layer_name)
                     .map_err(|e| SinkError::MvtWriter(format!("{:?}", e)))?;
-                let buffer = self.buffer.entry((output, layer_name)).or_default();
-                buffer.push(feature);
+                if !self.buffer.contains_key(&(
+                    output.clone(),
+                    layer_name.clone(),
+                    compress_output.clone(),
+                )) {
+                    let result = self.flush_buffer(context)?;
+                    self.buffer.clear();
+                    self.join_handles.extend(result);
+                }
+                let buffer = self
+                    .buffer
+                    .entry((output, layer_name, compress_output))
+                    .or_default();
+                buffer.push(feature.clone());
             }
             _ => {
                 return Err(Box::new(SinkError::MvtWriter(
@@ -180,442 +201,203 @@ impl Sink for MVTWriter {
         Ok(())
     }
     fn finish(&self, ctx: NodeContext) -> Result<(), BoxedError> {
-        for ((output, layer_name), buffer) in &self.buffer {
-            self.write(ctx.as_context(), buffer, output, layer_name)?;
+        let result = self.flush_buffer(ctx.as_context())?;
+        let mut join_handles = self.join_handles.clone();
+        join_handles.extend(result);
+
+        let timeout = std::time::Duration::from_secs(60 * 10);
+        let mut errors = Vec::new();
+
+        for (i, join) in join_handles.iter().enumerate() {
+            match join.lock().recv_timeout(timeout) {
+                Ok(_) => continue,
+                Err(mpsc::RecvTimeoutError::Timeout) => {
+                    errors.push(format!("Worker thread {} timed out after {:?}", i, timeout));
+                }
+                Err(mpsc::RecvTimeoutError::Disconnected) => {
+                    ctx.event_hub.warn_log(
+                        None,
+                        format!("Worker thread {} disconnected unexpectedly", i),
+                    );
+                }
+            }
+        }
+        if !errors.is_empty() {
+            return Err(SinkError::MvtWriter(format!(
+                "Failed to complete all worker threads: {}",
+                errors.join("; ")
+            ))
+            .into());
         }
         Ok(())
     }
 }
 
 impl MVTWriter {
+    #[allow(clippy::type_complexity)]
+    pub(crate) fn flush_buffer(&self, ctx: Context) -> crate::errors::Result<Vec<JoinHandle>> {
+        let mut result = Vec::new();
+        let mut features = HashMap::<(Uri, Option<Uri>, String), Vec<Feature>>::new();
+        for ((output, layer_name, compress_output), buffer) in &self.buffer {
+            features
+                .entry((output.clone(), compress_output.clone(), layer_name.clone()))
+                .or_default()
+                .extend(buffer.clone());
+        }
+        for ((output, compress_output, layer_name), buffer) in &features {
+            let res = self.write(
+                ctx.clone(),
+                buffer.clone(),
+                output,
+                layer_name,
+                compress_output,
+            )?;
+            result.extend(res);
+        }
+        Ok(result)
+    }
+
     pub fn write(
         &self,
         ctx: Context,
-        upstream: &[Feature],
+        upstream: Vec<Feature>,
         output: &Uri,
         layer_name: &str,
-    ) -> crate::errors::Result<()> {
+        compress_output: &Option<Uri>,
+    ) -> crate::errors::Result<Vec<JoinHandle>> {
         let tile_id_conv = TileIdMethod::Hilbert;
         let name = self.name().to_string();
-        std::thread::scope(|scope| {
-            let (sender_sliced, receiver_sliced) = std::sync::mpsc::sync_channel(2000);
-            let (sender_sorted, receiver_sorted) = std::sync::mpsc::sync_channel(2000);
-            scope.spawn(|| {
-                let result = geometry_slicing_stage(
-                    ctx.clone(),
-                    upstream,
-                    tile_id_conv,
-                    sender_sliced,
-                    output,
-                    layer_name,
-                    self.params.min_zoom,
-                    self.params.max_zoom,
+        let (sender_sliced, receiver_sliced) = std::sync::mpsc::sync_channel(2000);
+        let (sender_sorted, receiver_sorted) = std::sync::mpsc::sync_channel(2000);
+        let min_zoom = self.params.min_zoom;
+        let max_zoom = self.params.max_zoom;
+        let gctx = ctx.clone();
+        let out = output.clone();
+        let layer_name = layer_name.to_string();
+
+        let mut result = Vec::new();
+
+        let (tx, rx) = std::sync::mpsc::channel();
+        result.push(Arc::new(parking_lot::Mutex::new(rx)));
+        std::thread::spawn(move || {
+            let result = super::pipeline::geometry_slicing_stage(
+                gctx.clone(),
+                &upstream,
+                tile_id_conv,
+                sender_sliced,
+                &out,
+                &layer_name,
+                min_zoom,
+                max_zoom,
+            );
+            if let Err(err) = &result {
+                gctx.event_hub.error_log(
+                    None,
+                    format!("Failed to geometry_slicing_stage with error =  {:?}", err),
                 );
-                if let Err(err) = result {
-                    ctx.event_hub.error_log(
+                gctx.event_hub
+                    .send(Event::SinkFinishFailed { name: name.clone() });
+            }
+            tx.send(result).unwrap();
+        });
+        let name = self.name().to_string();
+        let gctx = ctx.clone();
+        let (tx, rx) = std::sync::mpsc::channel();
+        result.push(Arc::new(parking_lot::Mutex::new(rx)));
+        std::thread::spawn(move || {
+            let result = super::pipeline::feature_sorting_stage(receiver_sliced, sender_sorted);
+            if let Err(err) = &result {
+                ctx.event_hub.error_log(
+                    None,
+                    format!("Failed to feature_sorting_stage with error =  {:?}", err),
+                );
+                ctx.event_hub
+                    .send(Event::SinkFinishFailed { name: name.clone() });
+            }
+            tx.send(result).unwrap();
+        });
+        let out = output.clone();
+        let gctx = gctx.clone();
+        let name = self.name().to_string();
+        let compress_output = compress_output.clone();
+        let (tx, rx) = std::sync::mpsc::channel();
+        result.push(Arc::new(parking_lot::Mutex::new(rx)));
+        std::thread::spawn(move || {
+            let pool = rayon::ThreadPoolBuilder::new()
+                .use_current_thread()
+                .build()
+                .unwrap();
+            pool.install(|| {
+                let result = super::pipeline::tile_writing_stage(
+                    gctx.clone(),
+                    &out,
+                    receiver_sorted,
+                    tile_id_conv,
+                );
+                if let Err(err) = &result {
+                    gctx.event_hub.error_log(
                         None,
-                        format!("Failed to geometry_slicing_stage with error =  {:?}", err),
+                        format!("Failed to tile_writing_stage with error =  {:?}", err),
                     );
-                    ctx.event_hub
+                    gctx.event_hub
                         .send(Event::SinkFinishFailed { name: name.clone() });
                 }
-            });
-            scope.spawn(|| {
-                let result = feature_sorting_stage(receiver_sliced, sender_sorted);
-                if let Err(err) = result {
-                    ctx.event_hub.error_log(
-                        None,
-                        format!("Failed to feature_sorting_stage with error =  {:?}", err),
-                    );
-                    ctx.event_hub
-                        .send(Event::SinkFinishFailed { name: name.clone() });
-                }
-            });
-            scope.spawn(|| {
-                let pool = rayon::ThreadPoolBuilder::new()
-                    .use_current_thread()
-                    .build()
-                    .unwrap();
-                pool.install(|| {
-                    let result =
-                        tile_writing_stage(ctx.clone(), output, receiver_sorted, tile_id_conv);
-                    if let Err(err) = result {
-                        ctx.event_hub.error_log(
-                            None,
-                            format!("Failed to tile_writing_stage with error =  {:?}", err),
-                        );
-                        ctx.event_hub
-                            .send(Event::SinkFinishFailed { name: name.clone() });
-                    }
-                })
-            });
-        });
-        Ok(())
-    }
-}
 
-#[allow(clippy::too_many_arguments)]
-fn geometry_slicing_stage(
-    ctx: Context,
-    upstream: &[Feature],
-    tile_id_conv: TileIdMethod,
-    sender_sliced: std::sync::mpsc::SyncSender<(u64, Vec<u8>)>,
-    output_path: &Uri,
-    layer_name: &str,
-    min_zoom: u8,
-    max_zoom: u8,
-) -> crate::errors::Result<()> {
-    let bincode_config = bincode::config::standard();
-    let tile_contents = Arc::new(Mutex::new(Vec::new()));
-    let storage = ctx
-        .storage_resolver
-        .resolve(output_path)
-        .map_err(|e| crate::errors::SinkError::MvtWriter(format!("{:?}", e)))?;
-
-    // Convert CityObjects to sliced features
-    upstream.iter().par_bridge().try_for_each(|feature| {
-        let max_detail = 12; // 4096
-        let buffer_pixels = 5;
-        let tile_content = slice_cityobj_geoms(
-            feature,
-            layer_name,
-            min_zoom,
-            max_zoom,
-            max_detail,
-            buffer_pixels,
-            |(z, x, y, typename), mpoly| {
-                let feature = super::slice::SlicedFeature {
-                    typename,
-                    multi_polygons: mpoly,
-                    multi_line_strings: MultiLineString2::new(),
-                    properties: feature.attributes.clone(),
-                };
-                let bytes =
-                    bincode::serde::encode_to_vec(&feature, bincode_config).map_err(|err| {
-                        crate::errors::SinkError::MvtWriter(format!(
-                            "Failed to serialize a sliced feature: {:?}",
-                            err
-                        ))
-                    })?;
-                let tile_id = tile_id_conv.zxy_to_id(z, x, y);
-                if sender_sliced.send((tile_id, bytes)).is_err() {
-                    return Err(crate::errors::SinkError::MvtWriter("Canceled".to_string()));
-                };
-                Ok(())
-            },
-            |(z, x, y, typename), line_strings| {
-                let feature = super::slice::SlicedFeature {
-                    typename,
-                    multi_polygons: MultiPolygon2::new(),
-                    multi_line_strings: line_strings,
-                    properties: feature.attributes.clone(),
-                };
-                let bytes =
-                    bincode::serde::encode_to_vec(&feature, bincode_config).map_err(|err| {
-                        crate::errors::SinkError::MvtWriter(format!(
-                            "Failed to serialize a sliced feature: {:?}",
-                            err
-                        ))
-                    })?;
-                let tile_id = tile_id_conv.zxy_to_id(z, x, y);
-                if sender_sliced.send((tile_id, bytes)).is_err() {
-                    return Err(crate::errors::SinkError::MvtWriter("Canceled".to_string()));
-                };
-                Ok(())
-            },
-        )?;
-        tile_contents
-            .lock()
-            .map_err(|e| crate::errors::SinkError::MvtWriter(format!("Mutex poisoned: {}", e)))?
-            .push(tile_content);
-        Ok(())
-    })?;
-
-    let mut tile_content = TileContent::default();
-    for content in tile_contents
-        .lock()
-        .map_err(|e| crate::errors::SinkError::MvtWriter(format!("Mutex poisoned: {}", e)))?
-        .iter()
-    {
-        tile_content.min_lng = tile_content.min_lng.min(content.min_lng);
-        tile_content.max_lng = tile_content.max_lng.max(content.max_lng);
-        tile_content.min_lat = tile_content.min_lat.min(content.min_lat);
-        tile_content.max_lat = tile_content.max_lat.max(content.max_lat);
-    }
-    let metadata = TileMetadata::from_tile_content(
-        layer_name.to_string(),
-        min_zoom,
-        max_zoom,
-        &TileContent {
-            min_lng: tile_content.min_lng,
-            max_lng: tile_content.max_lng,
-            min_lat: tile_content.min_lat,
-            max_lat: tile_content.max_lat,
-        },
-    );
-
-    serde_json::to_string_pretty(&metadata)
-        .map_err(|e| crate::errors::SinkError::MvtWriter(format!("{:?}", e)))
-        .and_then(|metadata| {
-            storage
-                .put_sync(
-                    &output_path
-                        .join(Path::new("metadata.json"))
-                        .map_err(|e| crate::errors::SinkError::MvtWriter(format!("{:?}", e)))?
-                        .path(),
-                    bytes::Bytes::from(metadata),
-                )
-                .map_err(|e| crate::errors::SinkError::MvtWriter(format!("{:?}", e)))
-        })?;
-
-    Ok(())
-}
-
-fn feature_sorting_stage(
-    receiver_sliced: std::sync::mpsc::Receiver<(u64, Vec<u8>)>,
-    sender_sorted: std::sync::mpsc::SyncSender<(u64, Vec<Vec<u8>>)>,
-) -> crate::errors::Result<()> {
-    let config = kv_extsort::SortConfig::default().max_chunk_bytes(256 * 1024 * 1024);
-
-    let sorted_iter = kv_extsort::sort(
-        receiver_sliced
-            .into_iter()
-            .map(|(tile_id, body)| std::result::Result::<_, Infallible>::Ok((tile_id, body))),
-        config,
-    );
-
-    for ((_, tile_id), grouped) in &sorted_iter.chunk_by(|feat| match feat {
-        Ok((tile_id, _)) => (false, *tile_id),
-        Err(_) => (true, 0),
-    }) {
-        let grouped = grouped
-            .into_iter()
-            .map_ok(|(_, serialized_feats)| serialized_feats)
-            .collect::<kv_extsort::Result<Vec<_>, _>>();
-        match grouped {
-            Ok(serialized_feats) => {
-                if sender_sorted.send((tile_id, serialized_feats)).is_err() {
-                    return Err(crate::errors::SinkError::MvtWriter("Canceled".to_string()));
-                }
-            }
-            Err(kv_extsort::Error::Canceled) => {
-                return Err(crate::errors::SinkError::MvtWriter("Canceled".to_string()));
-            }
-            Err(err) => {
-                return Err(crate::errors::SinkError::MvtWriter(format!(
-                    "Failed to sort features: {:?}",
-                    err
-                )));
-            }
-        }
-    }
-
-    Ok(())
-}
-
-#[derive(Default)]
-struct LayerData {
-    pub features: Vec<vector_tile::tile::Feature>,
-    pub tags_enc: TagsEncoder,
-}
-
-fn tile_writing_stage(
-    ctx: Context,
-    output_path: &Uri,
-    receiver_sorted: std::sync::mpsc::Receiver<(u64, Vec<Vec<u8>>)>,
-    tile_id_conv: TileIdMethod,
-) -> crate::errors::Result<()> {
-    let default_detail = 12;
-    let min_detail = 9;
-
-    let storage = ctx
-        .storage_resolver
-        .resolve(output_path)
-        .map_err(|e| crate::errors::SinkError::MvtWriter(format!("{:?}", e)))?;
-
-    receiver_sorted
-        .into_iter()
-        .par_bridge()
-        .try_for_each(|(tile_id, serialized_feats)| {
-            let (zoom, x, y) = tile_id_conv.id_to_zxy(tile_id);
-
-            let path = output_path
-                .join(Path::new(&format!("{zoom}/{x}/{y}.pbf")))
-                .map_err(|e| crate::errors::SinkError::MvtWriter(format!("{:?}", e)))?;
-            for detail in (min_detail..=default_detail).rev() {
-                // Make a MVT tile binary
-                let bytes = make_tile(detail, &serialized_feats)?;
-
-                // Retry with a lower detail level if the compressed tile size is too large
-                let compressed_size = {
-                    let mut e = ZlibEncoder::new(Vec::new(), Compression::default());
-                    e.write_all(&bytes)
-                        .map_err(|e| crate::errors::SinkError::MvtWriter(format!("{:?}", e)))?;
-                    let compressed_bytes = e
-                        .finish()
-                        .map_err(|e| crate::errors::SinkError::MvtWriter(format!("{:?}", e)))?;
-                    compressed_bytes.len()
-                };
-                if detail != min_detail && compressed_size > 500_000 {
-                    // If the tile is too large, try a lower detail level
-                    continue;
-                }
-                storage
-                    .put_sync(&path.path(), bytes::Bytes::from(bytes))
-                    .map_err(|e| crate::errors::SinkError::MvtWriter(format!("{:?}", e)))?;
-                break;
-            }
-            Ok::<(), crate::errors::SinkError>(())
-        })?;
-    Ok(())
-}
-
-fn make_tile(default_detail: i32, serialized_feats: &[Vec<u8>]) -> crate::errors::Result<Vec<u8>> {
-    let mut layers: HashMap<String, LayerData> = HashMap::new();
-    let mut int_ring_buf = Vec::new();
-    let mut int_ring_buf2 = Vec::new();
-    let extent = 1 << default_detail;
-    let bincode_config = bincode::config::standard();
-
-    for serialized_feat in serialized_feats {
-        let (feature, _): (super::slice::SlicedFeature, _) =
-            bincode::serde::decode_from_slice(serialized_feat, bincode_config).map_err(|err| {
-                crate::errors::SinkError::MvtWriter(format!(
-                    "Failed to deserialize a sliced feature: {:?}",
-                    err
-                ))
-            })?;
-
-        let mpoly = feature.multi_polygons;
-        let mut int_mpoly = NMultiPolygon::<[i16; 2]>::new();
-
-        for poly in &mpoly {
-            for (ri, ring) in poly.rings().enumerate() {
-                int_ring_buf.clear();
-                int_ring_buf.extend(ring.into_iter().map(|[x, y]| {
-                    let x = (x * extent as f64 + 0.5) as i16;
-                    let y = (y * extent as f64 + 0.5) as i16;
-                    [x, y]
-                }));
-
-                // some simplification
-                {
-                    int_ring_buf2.clear();
-                    int_ring_buf2.push(int_ring_buf[0]);
-                    for c in int_ring_buf.windows(3) {
-                        let &[prev, curr, next] = c.try_into().map_err(|_| {
-                            crate::errors::SinkError::MvtWriter("Failed to convert".to_string())
-                        })?;
-
-                        // Remove duplicate points
-                        if prev == curr {
-                            continue;
+                if let Some(compress_output) = compress_output {
+                    if let Ok(storage) = gctx.storage_resolver.resolve(&compress_output) {
+                        let buffer = Vec::new();
+                        let mut cursor = Cursor::new(buffer);
+                        let writer = BufWriter::new(&mut cursor);
+                        let zip_result = reearth_flow_common::zip::write(
+                            writer,
+                            out.path().as_path(),
+                            out.path().as_path(),
+                        )
+                        .map_err(|e| crate::errors::SinkError::cesium3dtiles_writer(e.to_string()));
+                        match zip_result {
+                            Ok(_) => {
+                                match storage
+                                    .put_sync(
+                                        compress_output.path().as_path(),
+                                        bytes::Bytes::from(cursor.into_inner()),
+                                    )
+                                    .map_err(crate::errors::SinkError::cesium3dtiles_writer)
+                                {
+                                    Ok(_) => match std::fs::remove_dir_all(out.path().as_path()) {
+                                        Ok(_) => {}
+                                        Err(e) => {
+                                            gctx.event_hub.error_log(
+                                                None,
+                                                format!(
+                                                    "Failed to remove directory with error = {:?}",
+                                                    e
+                                                ),
+                                            );
+                                        }
+                                    },
+                                    Err(e) => {
+                                        gctx.event_hub.error_log(
+                                            None,
+                                            format!(
+                                                "Failed to write zip file with error = {:?}",
+                                                e
+                                            ),
+                                        );
+                                    }
+                                }
+                            }
+                            Err(e) => {
+                                gctx.event_hub.error_log(
+                                    None,
+                                    format!("Failed to write zip file with error = {:?}", e),
+                                );
+                            }
                         }
-
-                        // Reject collinear points
-                        let [curr_x, curr_y] = curr;
-                        let [prev_x, prev_y] = prev;
-                        let [next_x, next_y] = next;
-                        if curr != next
-                            && ((next_y - prev_y) as i32 * (curr_x - prev_x) as i32).abs()
-                                == ((curr_y - prev_y) as i32 * (next_x - prev_x) as i32).abs()
-                        {
-                            continue;
-                        }
-
-                        int_ring_buf2.push(curr);
-                    }
-                    int_ring_buf2.push(*int_ring_buf.last().ok_or(
-                        crate::errors::SinkError::MvtWriter("Failed to get last".to_string()),
-                    )?);
-                }
-
-                match ri {
-                    0 => int_mpoly.add_exterior(int_ring_buf2.drain(..)),
-                    _ => int_mpoly.add_interior(int_ring_buf2.drain(..)),
-                }
-            }
-        }
-
-        let mut int_line_string = NMultiLineString::<[i16; 2]>::new();
-        let mline_string = feature.multi_line_strings;
-
-        let mut int_line_string_buf = Vec::new();
-        for line_string in &mline_string {
-            int_line_string_buf.clear();
-            int_line_string_buf.extend(line_string.into_iter().map(|[x, y]| {
-                let x = (x * extent as f64 + 0.5) as i16;
-                let y = (y * extent as f64 + 0.5) as i16;
-                [x, y]
-            }));
-            int_line_string.add_linestring(&LineString2::from_raw(
-                int_line_string_buf.drain(..).collect(),
-            ));
-        }
-
-        // encode geometry
-        let mut geom_enc = GeometryEncoder::new();
-        for poly in &int_mpoly {
-            let exterior = poly.exterior();
-            if exterior.signed_ring_area() > 0.0 {
-                geom_enc.add_ring(&exterior);
-                for interior in poly.interiors() {
-                    if interior.is_cw() {
-                        geom_enc.add_ring(&interior);
                     }
                 }
-            }
-        }
-
-        for line_string in &int_line_string {
-            let area = line_string.signed_ring_area();
-            if area as i32 != 0 {
-                geom_enc.add_linestring(&line_string);
-            }
-        }
-        let geometry = geom_enc.into_vec();
-        if geometry.is_empty() {
-            continue;
-        }
-
-        let layer = {
-            let layer = layers.entry(feature.typename).or_default();
-
-            // Encode attributes as MVT tags
-            for (key, value) in &feature.properties {
-                convert_properties(&mut layer.tags_enc, key.as_ref(), value);
-            }
-            layer
-        };
-
-        layer.features.push(vector_tile::tile::Feature {
-            id: None,
-            tags: layer.tags_enc.take_tags(),
-            r#type: Some(vector_tile::tile::GeomType::Polygon as i32),
-            geometry,
-        });
-    }
-
-    let layers = layers
-        .into_iter()
-        .flat_map(|(name, layer_data)| {
-            if layer_data.features.is_empty() {
-                return None;
-            }
-            let (keys, values) = layer_data.tags_enc.into_keys_and_values();
-            Some(vector_tile::tile::Layer {
-                version: 2,
-                name: name.to_string(),
-                features: layer_data.features,
-                keys,
-                values,
-                extent: Some(extent),
+                tx.send(result).unwrap();
             })
-        })
-        .collect();
-
-    let tile = vector_tile::Tile { layers };
-
-    let bytes = tile.encode_to_vec();
-    Ok(bytes)
+        });
+        Ok(result)
+    }
 }
