+++ engine/schema/actions_es.json
@@ -2549,6 +2549,16 @@
           "output"
         ],
         "properties": {
+          "compressOutput": {
+            "anyOf": [
+              {
+                "$ref": "#/definitions/Expr"
+              },
+              {
+                "type": "null"
+              }
+            ]
+          },
           "layerName": {
             "$ref": "#/definitions/Expr"
           },
